A4/backend/requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
pandas
matplotlib
